+++ data-gov-ckan/src/models/tag.rs
@@ -38,7 +38,9 @@ impl Tag {
         }
     }
 }
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
+#[derive(
+    Clone, Copy, Debug, Default, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize,
+)]
 pub enum State {
     #[serde(rename = "active")]
     #[default]