+++ data-gov-ckan/src/models/user.rs
@@ -65,7 +65,9 @@ impl User {
         }
     }
 }
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
+#[derive(
+    Clone, Copy, Debug, Default, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize,
+)]
 pub enum State {
     #[serde(rename = "active")]
     #[default]